requirements.txt
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
dateparser==1.0.0
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
GoogleNews==1.5.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
pandas==1.1.0
plotly==4.14.3
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
soupsieve==2.1
git+git://github.com/areed1192/td-ameritrade-python-api.git
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
